Move some functions to python/common.py
