+++ python/common.py
@@ -12,6 +12,8 @@ import tempfile
 import subprocess
 import numpy as np
 
+
+from python.rpc_utils import corresponding_roi
 from config import cfg
 
 
@@ -385,6 +387,74 @@ def image_zoom_gdal(im, f, out=None, w=None, h=None):
          ' %d %d %s %s') % (w/float(f), h/float(f), tmp, out))
     return out
 
+
+
+def cropImage(inp,out,col,row,tw,th,z):
+    """
+    Extract an ROI from inp to out, with ROI defined as (col,row,tw,th) <==> (Upper left corner, size of ROI), and apply a zoom z.
+    """
+    
+    if z == 1:
+        image_crop_TIFF(inp, col,row,tw,th, out)
+    else:
+        # gdal is used for the zoom because it handles BigTIFF files, and
+        # before the zoom out the image may be that big
+        tmp_crop = image_crop_TIFF(inp, col,row,tw,th)
+        image_zoom_gdal(tmp_crop, z, out, tw,th)
+        
+
+
+def getMinMaxFromExtract(tile_dir,tilesFullInfo):
+    """
+    Get min/max intensities of an extract ROI from the ref image.
+    
+    Args:
+        tile_dir : a key for the dictionnary tilesFullInfo; refers to a particular tile
+        tilesFullInfo : a dictionnary that provides all you need to process a tile -> col,row,tw,th,ov,i,j,pos,images
+    """
+    
+    print "\nCrop ref image and compute min/max intensities..."
+
+    #Get info
+    col,row,tw,th,ov,i,j,pos,images=tilesFullInfo[tile_dir]
+    img1 = images[0]['img']
+    
+    # output files
+    crop_ref = tile_dir + '/roi_ref.tif'
+    local_minmax = tile_dir + '/local_minmax.txt'
+
+    z = cfg['subsampling_factor']
+    cropImage(img1,crop_ref,col,row,tw,th,z)
+		
+    image_getminmax(crop_ref,local_minmax)
+    
+    
+    
+def crop_corresponding_areas(out_dir, images, roi, zoom=1):
+    """
+    Crops areas corresponding to the reference ROI in the secondary images.
+
+    Args:
+        out_dir:
+        images: sequence of dicts containing the paths to input data
+        roi: dictionary containing the ROI definition
+        zoom: integer zoom out factor
+    """
+    rpc_ref = images[0]['rpc']
+    for i, image in enumerate(images[1:]):
+        x, y, w, h = corresponding_roi(rpc_ref, image['rpc'],
+                                                 roi['x'], roi['y'], roi['w'],
+                                                 roi['h'])
+        if zoom == 1:
+            image_crop_TIFF(image['img'], x, y, w, h,
+                                   '%s/roi_sec_%d.tif' % (out_dir, i))
+        else:
+            # gdal is used for the zoom because it handles BigTIFF files, and
+            # before the zoom out the image may be that big
+            tmp = image_crop_TIFF(image['img'], x, y, w, h)
+            image_zoom_gdal(tmp, zoom, '%s/roi_sec_%d.tif' % (out_dir,
+                                                                     i), w, h)    
+    
 
 def image_zoom_out_morpho(im, f):
     """
