+++ test.json
@@ -37,4 +37,5 @@
   "mosaic_method" : "piio",
   "msk_erosion": 0,
   "dsm_resolution": 0.5
+  "max_nb_threads": 3
 }
